Code review comments fixing
> remove the separated listener class and merge it to AlterSuiteListener.
> change XRay Integration class name.
> change cucumber report path.
> update XRay Integration class access modifiers.
