+++ src/main/java/com/shaft/tools/tms/XrayIntegrationHelper.java
@@ -4,6 +4,7 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
+import com.shaft.cli.FileActions;
 import com.shaft.tools.io.ReportManager;
 import com.shaft.tools.io.ReportManagerHelper;
 import io.restassured.config.RestAssuredConfig;
@@ -11,39 +12,45 @@ import io.restassured.config.SSLConfig;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
 import java.io.File;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.text.SimpleDateFormat;
 import java.util.Base64;
 import java.util.Calendar;
-
 import static io.restassured.RestAssured.*;
 import static io.restassured.config.EncoderConfig.encoderConfig;
 
-public class XrayIntegrationManager {
-     public static String _JiraAuthorization=
-             Base64.getEncoder().encodeToString(System.getProperty("authorization").getBytes());
-     public static String _ProjectKey= System.getProperty("projectKey");
-     public static String _TestExecutionID=null;
 
-     public static void setup() {
+public class XrayIntegrationHelper {
+
+     private static final String _JiraAuthorization =
+                Base64.getEncoder().encodeToString(System.getProperty("authorization").trim().getBytes());
+     private static final String _ProjectKey = System.getProperty("projectKey").trim();
+     private static String _TestExecutionID = null;
+
+     private static void setup() {
          baseURI = System.getProperty("jiraUrl");
          given()
                  .config(RestAssuredConfig.config().sslConfig(SSLConfig.sslConfig().allowAllHostnames()))
                  .relaxedHTTPSValidation();
      }
 
+    /**
+     * Import cucumber results recorded in cucumber.jsom file through /import/execution/cucumber endpoint
+     * @param filepath > the report relative path
+     *
+     */
     public static void importCucumberResults(String filepath) throws Exception {
 
         setup();
-        String reportPath = System.getProperty("user.dir")+"\\"+filepath;
+        String reportPath = FileActions.getAbsolutePath(filepath);
         ReportManager.logDiscrete("uploading file: "+reportPath);
         ReportManager.logDiscrete("Length: "+new File(reportPath).length());
 
         Gson gson = new GsonBuilder().setPrettyPrinting().create();
         JsonElement je = JsonParser.parseString( new String(Files.readAllBytes(Paths.get(reportPath))));
         String prettyJsonString = gson.toJson(je);
-        ReportManager.logDiscrete("Pretty: \n"+ prettyJsonString);
 
         try {
             Response response =given()
@@ -53,7 +60,7 @@ public class XrayIntegrationManager {
                     .expect().statusCode(200)
                     .when()
                     .post( "/rest/raven/1.0/import/execution/cucumber").then().extract().response();
-            ReportManager.logDiscrete("#################"+ response.asString());
+
             _TestExecutionID=response.jsonPath().get("testExecIssue.key").toString();
             ReportManager.logDiscrete("ExecutionID: "+_TestExecutionID);
         }catch (Exception e){
@@ -61,35 +68,46 @@ public class XrayIntegrationManager {
         }
     }
 
+    /**
+     * Import TestNG results recorded in testng-results.jsom file through /import/execution/testng?projectKey=[projectKey] endpoint
+     * @param  executionName  > The execution name mentioned in JiraXray.properties
+     * @param  executionDescription > The execution Description mentioned in JiraXray.properties
+     */
     public static void renameTestExecutionSuit(String executionName, String executionDescription) {
+
          if (_TestExecutionID==null) return;
          setup();
-        Calendar cal = Calendar.getInstance();
-        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
+         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 
          String body= "{\r\n    \"fields\" : {\r\n       " +
                  " \"summary\": " +
-                 "\"Execution results "+executionName+" | "+sdf.format(cal.getTime())+"\",\r\n        " +
+                 "\"Execution results "+executionName+" | "+sdf.format(Calendar.getInstance().getTime())+"\",\r\n        " +
                  "\"description\": " +
                  "\""+executionDescription+"\"\r\n    }\r\n}";
         try {
-            Response response =given()
+            given()
                     .contentType("application/json")
                     .header("Authorization", "Basic " + _JiraAuthorization)
                     .body(body)
                     .expect().statusCode(204)
                     .when()
                     .put( "/rest/api/2/issue/" +_TestExecutionID).then().extract().response();
-            ReportManager.logDiscrete("#################"+ response.asString());
 
         }catch (Exception e){
             ReportManagerHelper.log(e);
         }
     }
 
-    public static void importTestNGResults(String filepath) {
+    /**
+     * Import TestNG results recorded in testng-results.jsom file through /import/execution/testng?projectKey=[projectKey] endpoint
+     * @param filepath > the report relative path
+     *
+     */
+    public static void importTestNGResults(String filepath) throws IOException {
         setup();
-        String reportPath = System.getProperty("user.dir")+"\\"+filepath;
+        String reportPath = FileActions.getAbsolutePath(filepath);
+        ReportManager.logDiscrete("uploading file: "+reportPath);
+        ReportManager.logDiscrete("Length: "+new File(reportPath).length());
         try {
             Response response = given()
                     .config(config().encoderConfig(encoderConfig().encodeContentTypeAs("multipart/form-data", ContentType.TEXT)))
@@ -100,7 +118,6 @@ public class XrayIntegrationManager {
                     .post("/rest/raven/1.0/import/execution/testng?projectKey="+_ProjectKey)
                     .then().log().all().extract().response();
 
-            ReportManager.logDiscrete("#################"+ response.asString());
             _TestExecutionID=response.jsonPath().get("testExecIssue.key").toString();
             ReportManager.logDiscrete("ExecutionID: "+_TestExecutionID);
