+++ src/main/resources/defaultProperties/JiraXRay.properties
@@ -12,8 +12,8 @@ projectKey= <JiraProjectID: mandatory>
 authorization= <username>:<password>
 ###################################################
 # Report path can be either one of the following: #
-# target/cucumber/Destination/cucumber.json       #
-# target/surefire-reports/testng-results.xml      #
+# allure-results/cucumber.json                   #
+# target/surefire-reports/testng-results.xml     #
 ###################################################
 reportPath= <reportPath: mandatory>
 # <YourExecutionName: Optional>
