+++ src/main/resources/defaultProperties/cucumber.properties
@@ -1,5 +1,5 @@
 cucumber.filter.tags=
 cucumber.features=src/test/resources
 cucumber.extraGlue=
-cucumber.plugin=pretty, json:target/cucumber/Destination/cucumber.json, html:allure-results/cucumberReport.html, com.shaft.tools.listeners.CucumberFeatureListener
+cucumber.plugin=pretty, json:allure-results/cucumber.json, html:allure-results/cucumberReport.html, com.shaft.tools.listeners.CucumberFeatureListener
 cucumber.glue=com.shaft.cucumber
